Moved vertex package inside model package
